Update renv.lock
the R package class has a recommended version of >3,x,x but the current verison of r is 4,2,0. This collides with version management.

In the .lock file however there is no collision with these versions because class was set to the older version which still works.

However, the install.packages() function will ignore your version argument, so you'll get the latest version, 7.3-22. Version 7.3-20 is not compatible with R 4.3.x, because the Sint and related definitions have been dropped.